refactor(controller): isolate provider selection algorithm
